server/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
asgiref==3.11.0
    # via
    #   django
    #   django-cors-headers
beautifulsoup4==4.14.3
    # via yfinance
certifi==2025.11.12
    # via
    #   curl-cffi
    #   requests
cffi==2.0.0
    # via curl-cffi
charset-normalizer==3.4.4
    # via requests
contourpy==1.3.3
    # via matplotlib
curl-cffi==0.14.0
    # via yfinance
cycler==0.12.1
    # via matplotlib
django==5.2.9
    # via
    #   stock-predictor (pyproject.toml)
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.9.0
    # via stock-predictor (pyproject.toml)
djangorestframework==3.16.1
    # via stock-predictor (pyproject.toml)
filelock==3.20.1
    # via torch
fonttools==4.61.1
    # via matplotlib
frozendict==2.4.7
    # via yfinance
fsspec==2025.12.0
    # via torch
gunicorn==23.0.0
    # via stock-predictor (pyproject.toml)
idna==3.11
    # via requests
jinja2==3.1.6
    # via torch
joblib==1.5.3
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via
    #   stock-predictor (pyproject.toml)
    #   seaborn
mpmath==1.3.0
    # via sympy
multitasking==0.0.12
    # via yfinance
networkx==3.6.1
    # via torch
numpy==2.4.0
    # via
    #   stock-predictor (pyproject.toml)
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   torchvision
    #   yfinance
packaging==25.0
    # via
    #   gunicorn
    #   matplotlib
pandas==2.3.3
    # via
    #   stock-predictor (pyproject.toml)
    #   seaborn
    #   yfinance
peewee==3.18.3
    # via yfinance
pillow==12.0.0
    # via
    #   matplotlib
    #   torchvision
platformdirs==4.5.1
    # via yfinance
protobuf==6.33.2
    # via yfinance
pycparser==2.23
    # via cffi
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via
    #   pandas
    #   yfinance
pyyaml==6.0.3
    # via stock-predictor (pyproject.toml)
requests==2.32.5
    # via yfinance
scikit-learn==1.8.0
    # via stock-predictor (pyproject.toml)
scipy==1.16.3
    # via scikit-learn
seaborn==0.13.2
    # via stock-predictor (pyproject.toml)
setuptools==80.9.0
    # via torch
six==1.17.0
    # via python-dateutil
soupsieve==2.8.1
    # via beautifulsoup4
sqlparse==0.5.5
    # via django
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
torch==2.9.1
    # via
    #   stock-predictor (pyproject.toml)
    #   torchvision
torchvision==0.24.1
    # via stock-predictor (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   torch
tzdata==2025.3
    # via
    #   django
    #   pandas
urllib3==2.6.2
    # via requests
websockets==15.0.1
    # via yfinance
yfinance==0.2.66
    # via stock-predictor (pyproject.toml)
